Enhance SpMV implementation and build process
- Updated compile.sh to include CUDA and MKL support during the build process.
- Refactored spmv_kernels.cu to improve kernel structure and added checks for zero rows.
- Modified pi_csr.cpp to change data types from size_t to int for row_ptr and col_idx, ensuring compatibility with CUDA.
- Enhanced error handling in csr_from_bin function to prevent overflow and invalid CSR structures.
- Updated pi_spmv.cpp to use int instead of size_t for indexing, improving consistency across the codebase.
- Removed redundant MIN and MAX macros from utils.h, utilizing std::min and std::max instead.
- Improved test_spmv.cpp to handle both single files and directories for input, enhancing testing flexibility.
- Added utility functions for directory and file handling in test_spmv.cpp.

+++ src/pi_csr.h
@@ -3,14 +3,15 @@
 
 typedef struct
 {
-    size_t n_rows;
-    size_t n_cols;
-    size_t nnz;      // 非零元素个数
-    size_t *row_ptr; // 长度为n_rows + 1
-    size_t *col_idx;
+    int n_rows;
+    int n_cols;
+    int nnz;
+    int *row_ptr;
+    int *col_idx;
     double *values;
 } pi_csr;
-piState csr_create(size_t n_rows, size_t n_cols, size_t nnz, pi_csr *dist);
+
+piState csr_create(int n_rows, int n_cols, int nnz, pi_csr *dist);
 void csr_destroy(pi_csr *A);
 void csr_print(const pi_csr *A);
 piState csr_from_bin(const char *src_file, pi_csr *dist);
